requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
scipy
fpdf2
matplotlib
seaborn
faker
groq